requirement.txt
fastapi>=0.110
uvicorn>=0.23

pydantic>=2.7

langchain>=0.3
langchain-openai>=0.1
langchain-community>=0.3

faiss-cpu>=1.7
tiktoken>=0.7
python-dotenv>=1.0


